Created SolarTrough separate file and also added relevant test for ew and ns.
